Criada a funcionalidade que passa o ID do usuário para as outras telas

+++ todolist/src/main/java/br/com/best/model/Usuario.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public class Usuario {
 
-    private Long userId;
+    private int userId;
     private String nickName;
     private String name;
     private String password;
@@ -13,7 +13,7 @@ public class Usuario {
     public Usuario() {
     }
 
-    public Usuario(Long userId, String nickName, String name, String password, LocalDateTime createdAt) {
+    public Usuario(int userId, String nickName, String name, String password, LocalDateTime createdAt) {
         this.userId = userId;
         this.nickName = nickName;
         this.name = name;
@@ -21,11 +21,11 @@ public class Usuario {
         this.createdAt = createdAt;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
